Add files via upload
Who the audience of the dashboard is:
The logistics dashboard is used by logistics managers to help them understand the flow of goods. The report also provides a view by product, delivery method, and location for further insights.
What the purpose of the dashboard is:
The purpose of the logistics dashboard is to help logistics managers with product flow. The dashboard identifies areas of high/low product flow to helps managers allocate resources in order to meet delivery standards and run a smooth operation. Overall, this helps the organisation become more efficient by better utilising their resources, essentially saving them time and money.

What the dashboards key insights are:
1.	KPI's – Numbers of Order, Numbers of Quantity, Numbers of Customers:
To provide context relating to the scale of the operation.
2.	Decomposition Tree - Flow of orders by category and sub-category:
To prioritize each products shipping process.
3.	Fill Map - Orders by state:
Helps managers allocate resources based on demand.
4.	Treemap - Orders by date:
To have vivid understanding of daily, weekly, monthly, quarterly and yearly numbers of Orders. This can also be good for forecasting.