Add files via upload
Seoul Bike Rental Dataset
This is a popular dataset which I got from https://www.kaggle.com/datasets, I have downloaded tons of datasets from this site for over a year. I decided to play with different colour combinations and it’s amazing.
I have done this project using R-programming language and Excel and both can be viewed on my GitHub page.
It’s a simple dataset that shows the numbers of bikes rented in a given period and conditions that might aid renting of bike or prevent it.